Update README.md Step 5
Step 5:
You are about to do something that you should never do again. Go to the Code tab of your repository (on GitHub). Find and click on the pencil icon. Edit the file. Scroll to the bottom and click the Commit button. We will be committing directly to the main/master
branch so make sure that option is selected.